+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Fixed the T2CL CPU template to pass through the RSBA and RRSBA bits of the
   IA32_ARCH_CAPABILITIES MSR from the host in accordance with an Intel microcode
   update.
+- Fixed passing through cache information from host in CPUID leaf 0x80000005.
 
 ## [1.3.0]